Fix failing tests
This isn't as good as actually fleshing these tests out, but at least it
gets them passing in CI.

The problem was that unit tests need things it depends on from the
container to be explicitly stated, but I didn't configure that when
I set up `ember-simple-auth`. Telling the tests that they need to
register the `session` service was enough to get the basic tests passing
again.